requirements.txt
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2022.12.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
exceptiongroup==1.1.1
fastapi==0.94.1
flake8==6.0.0
flake8-black==0.3.6
flower==1.2.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanize==4.6.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
kombu==5.2.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pycodestyle==2.10.0
pydantic==1.10.6
pyflakes==3.0.1
pytest==7.2.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
# redis-server==6.0.9
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.6
SQLAlchemy-Utils==0.40.0
starlette==0.26.1
tomli==2.0.1
tornado==6.2
typing-extensions==4.5.0
uvicorn==0.21.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4